+++ Clockwork/DataSource/EloquentDataSource.php
@@ -28,17 +28,8 @@ class EloquentDataSource extends DataSource
 
 	// Query counts by type
 	protected $count = [
-		'total'  => 0,
-		'slow'   => 0,
-		'select' => 0,
-		'insert' => 0,
-		'update' => 0,
-		'delete' => 0,
-		'other'  => 0
+		'total' => 0, 'slow' => 0, 'select' => 0, 'insert' => 0, 'update' => 0, 'delete' => 0, 'other' => 0
 	];
-
-	// Array of filter functions for collected queries
-	protected $filters = [];
 
 	// Whether we are collecting cache queries or stats only
 	protected $collectQueries = true;
@@ -105,6 +96,7 @@ class EloquentDataSource extends DataSource
 			'query'      => $this->createRunnableQuery($event->sql, $event->bindings, $event->connectionName),
 			'duration'   => $event->time,
 			'connection' => $event->connectionName,
+			'time'       => microtime(true) - $event->time / 1000,
 			'trace'      => $shortTrace = (new Serializer)->trace($trace),
 			'file'       => isset($shortTrace[0]) ? $shortTrace[0]['file'] : null,
 			'line'       => isset($shortTrace[0]) ? $shortTrace[0]['line'] : null,
@@ -112,13 +104,15 @@ class EloquentDataSource extends DataSource
 			'tags'       => $this->slowThreshold !== null && $event->time > $this->slowThreshold ? [ 'slow' ] : []
 		];
 
+		$this->nextQueryModel = null;
+
+		if (! $this->passesFilters([ $query, $trace ], 'early')) return;
+
 		$this->incrementQueryCount($query);
 
-		if ($this->collectQueries && $this->passesFilters([ $query ])) {
-			$this->queries[] = $query;
-		}
+		if (! $this->collectQueries || ! $this->passesFilters([ $query, $trace ])) return;
 
-		$this->nextQueryModel = null;
+		$this->queries[] = $query;
 	}
 
 	/**
@@ -152,6 +146,18 @@ class EloquentDataSource extends DataSource
 		$this->appendDuplicateQueriesWarnings($request);
 
 		return $request;
+	}
+
+	// Reset the data source to an empty state, clearing any collected data
+	public function reset()
+	{
+		$this->queries = [];
+
+		$this->count = [
+			'total' => 0, 'slow' => 0, 'select' => 0, 'insert' => 0, 'update' => 0, 'delete' => 0, 'other' => 0
+		];
+
+		$this->nextQueryModel = null;
 	}
 
 	/**