+++ Clockwork/DataSource/LaravelEventsDataSource.php
@@ -49,6 +49,12 @@ class LaravelEventsDataSource extends DataSource
 		return $request;
 	}
 
+	// Reset the data source to an empty state, clearing any collected data
+	public function reset()
+	{
+		$this->events = [];
+	}
+
 	// Registers a new event, prepares data for serialization and resolves registered listeners
 	protected function registerEvent($event, array $data)
 	{
@@ -119,6 +125,7 @@ class LaravelEventsDataSource extends DataSource
 	{
 		return [
 			'Illuminate\\\\.+',
+			'Laravel\\\\.+',
 			'auth\.(?:attempt|login|logout)',
 			'artisan\.start',
 			'bootstrapped:.+',