+++ Clockwork/Request/RequestType.php
@@ -1,0 +1,10 @@
+<?php namespace Clockwork\Request;
+
+// Supported request types
+class RequestType
+{
+	const REQUEST = 'request';
+	const COMMAND = 'command';
+	const QUEUE_JOB = 'queue-job';
+	const TEST = 'test';
+}