+++ Clockwork/Support/Laravel/ClockworkServiceProvider.php
@@ -8,10 +8,15 @@ use Clockwork\DataSource\LaravelCacheDataSource;
 use Clockwork\DataSource\LaravelEventsDataSource;
 use Clockwork\DataSource\LaravelRedisDataSource;
 use Clockwork\DataSource\LaravelQueueDataSource;
+use Clockwork\DataSource\LaravelTwigDataSource;
+use Clockwork\DataSource\LaravelViewsDataSource;
 use Clockwork\DataSource\PhpDataSource;
 use Clockwork\DataSource\SwiftDataSource;
+use Clockwork\DataSource\TwigDataSource;
 use Clockwork\DataSource\XdebugDataSource;
+use Clockwork\Helpers\StackFilter;
 use Clockwork\Request\Log;
+use Clockwork\Request\Request;
 use Clockwork\Storage\StorageInterface;
 
 use Illuminate\Redis\RedisManager;
@@ -46,11 +51,21 @@ class ClockworkServiceProvider extends ServiceProvider
 		if ($support->isFeatureEnabled('cache')) $this->app['clockwork.cache']->listenToEvents();
 		if ($support->isFeatureEnabled('database')) $this->app['clockwork.eloquent']->listenToEvents();
 		if ($support->isFeatureEnabled('events')) $this->app['clockwork.events']->listenToEvents();
-		if ($support->isFeatureEnabled('queue')) $this->app['clockwork.queue']->listenToEvents();
+		if ($support->isFeatureEnabled('queue')) {
+			$this->app['clockwork.queue']->listenToEvents();
+			$this->app['clockwork.queue']->setCurrentRequestId($this->app['clockwork']->getRequest()->id);
+		}
 		if ($support->isFeatureEnabled('redis')) {
 			$this->app[RedisManager::class]->enableEvents();
 			$this->app['clockwork.redis']->listenToEvents();
 		}
+		if ($support->isFeatureEnabled('views')) {
+			$support->getConfig('features.views.use_twig_profiler', false)
+				? $this->app['clockwork.twig']->listenToEvents() : $this->app['clockwork.views']->listenToEvents();
+		}
+
+		if ($support->isCollectingCommands()) $support->collectCommands();
+		if ($support->isCollectingQueueJobs()) $support->collectQueueJobs();
 	}
 
 	protected function listenToFrameworkEvents()
@@ -69,6 +84,7 @@ class ClockworkServiceProvider extends ServiceProvider
 			$clockwork = (new Clockwork)
 				->setAuthenticator($app['clockwork.authenticator'])
 				->setLog($app['clockwork.log'])
+				->setRequest($app['clockwork.request'])
 				->setStorage($app['clockwork.storage'])
 				->addDataSource(new PhpDataSource())
 				->addDataSource($app['clockwork.laravel']);
@@ -80,6 +96,11 @@ class ClockworkServiceProvider extends ServiceProvider
 			if ($support->isFeatureEnabled('events')) $clockwork->addDataSource($app['clockwork.events']);
 			if ($support->isFeatureEnabled('emails')) $clockwork->addDataSource($app['clockwork.swift']);
 			if ($support->isFeatureAvailable('xdebug')) $clockwork->addDataSource($app['clockwork.xdebug']);
+			if ($support->isFeatureEnabled('views')) {
+				$clockwork->addDataSource(
+					$support->getConfig('features.views.use_twig_profiler', false) ? $app['clockwork.twig'] : $app['clockwork.views']
+				);
+			}
 
 			return $clockwork;
 		});
@@ -90,6 +111,10 @@ class ClockworkServiceProvider extends ServiceProvider
 
 		$this->app->singleton('clockwork.log', function ($app) {
 			return new Log;
+		});
+
+		$this->app->singleton('clockwork.request', function ($app) {
+			return new Request;
 		});
 
 		$this->app->singleton('clockwork.storage', function ($app) {
@@ -104,8 +129,8 @@ class ClockworkServiceProvider extends ServiceProvider
 		$this->registerDataSources();
 		$this->registerAliases();
 
+		$this->app->make('clockwork.request'); // instantiate the request to have id and time available as early as possible
 		$this->app['clockwork.support']->configureSerializer();
-
 		$this->app['clockwork.laravel']->listenToEarlyEvents();
 
 		if ($this->app['clockwork.support']->getConfig('register_helpers', true)) {
@@ -132,7 +157,7 @@ class ClockworkServiceProvider extends ServiceProvider
 		});
 
 		$this->app->singleton('clockwork.eloquent', function ($app) {
-			return (new EloquentDataSource(
+			$dataSource = (new EloquentDataSource(
 				$app['db'],
 				$app['events'],
 				$app['clockwork.support']->getConfig('features.database.collect_queries'),
@@ -140,6 +165,24 @@ class ClockworkServiceProvider extends ServiceProvider
 				$app['clockwork.support']->getConfig('features.database.slow_only'),
 				$app['clockwork.support']->getConfig('features.database.detect_duplicate_queries')
 			));
+
+			// if we are collecting queue jobs, filter out queries caused by the database queue implementation
+			if ($app['clockwork.support']->isCollectingQueueJobs()) {
+				$dataSource->addFilter(function ($query, $trace) {
+					return ! $trace->first(StackFilter::make()->isClass(\Illuminate\Queue\Worker::class));
+				}, 'early');
+			}
+
+			if ($app->runningUnitTests()) {
+				$dataSource->addFilter(function ($query, $trace) {
+					return ! $trace->first(StackFilter::make()->isClass([
+						\Illuminate\Database\Migrations\Migrator::class,
+						\Illuminate\Database\Console\Migrations\MigrateCommand::class
+					]));
+				});
+			}
+
+			return $dataSource;
 		});
 
 		$this->app->singleton('clockwork.events', function ($app) {
@@ -153,7 +196,6 @@ class ClockworkServiceProvider extends ServiceProvider
 			return (new LaravelDataSource(
 				$app,
 				$app['clockwork.support']->isFeatureEnabled('log'),
-				$app['clockwork.support']->isFeatureEnabled('views'),
 				$app['clockwork.support']->isFeatureEnabled('routes')
 			))
 				->setLog($app['clockwork.log']);
@@ -164,11 +206,36 @@ class ClockworkServiceProvider extends ServiceProvider
 		});
 
 		$this->app->singleton('clockwork.redis', function ($app) {
-			return (new LaravelRedisDataSource($app['events']));
+			$dataSource = new LaravelRedisDataSource($app['events']);
+
+			// if we are collecting queue jobs, filter out commands executed by the redis queue implementation
+			if ($app['clockwork.support']->isCollectingQueueJobs()) {
+				$dataSource->addFilter(function ($query, $trace) {
+					return ! $trace->first(StackFilter::make()->isClass([
+						\Illuminate\Queue\RedisQueue::class,
+						\Laravel\Horizon\Repositories\RedisJobRepository::class,
+						\Laravel\Horizon\Repositories\RedisTagRepository::class,
+						\Laravel\Horizon\Repositories\RedisMetricsRepository::class
+					]));
+				});
+			}
+
+			return $dataSource;
 		});
 
 		$this->app->singleton('clockwork.swift', function ($app) {
 			return new SwiftDataSource($app['mailer']->getSwiftMailer());
+		});
+
+		$this->app->singleton('clockwork.twig', function ($app) {
+			return new TwigDataSource($app['twig']);
+		});
+
+		$this->app->singleton('clockwork.views', function ($app) {
+			return new LaravelViewsDataSource(
+				$app['events'],
+				$app['clockwork.support']->getConfig('features.views.collect_data')
+			);
 		});
 
 		$this->app->singleton('clockwork.xdebug', function ($app) {