+++ Clockwork/Support/Laravel/ClockworkSupport.php
@@ -7,12 +7,14 @@ use Clockwork\Helpers\Serializer;
 use Clockwork\Helpers\ServerTiming;
 use Clockwork\Helpers\StackFilter;
 use Clockwork\Helpers\StackTrace;
+use Clockwork\Request\Request;
 use Clockwork\Storage\FileStorage;
 use Clockwork\Storage\Search;
 use Clockwork\Storage\SqlStorage;
 use Clockwork\Web\Web;
 
 use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\Console\Kernel as ConsoleKernel;
 use Illuminate\Http\JsonResponse;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -117,6 +119,108 @@ class ClockworkSupport
 		}
 	}
 
+	// Set up collecting of executed artisan commands
+	public function collectCommands()
+	{
+		$this->app['events']->listen(\Illuminate\Console\Events\CommandStarting::class, function ($event) {
+			// only collect commands ran through artisan cli, other commands are recorded as part of respective request
+			if (basename(StackTrace::get()->last()->file) != 'artisan') return;
+
+			if (! $this->getConfig('artisan.collect_output')) return;
+			if (! $event->command || $this->isCommandFiltered($event->command)) return;
+
+			$event->output->setFormatter(
+				new Console\CapturingFormatter($event->output->getFormatter())
+			);
+		});
+
+		$this->app['events']->listen(\Illuminate\Console\Events\CommandFinished::class, function ($event) {
+			// only collect commands ran through artisan cli, other commands are recorded as part of respective request
+			if (basename(StackTrace::get()->last()->file) != 'artisan') return;
+
+			if (! $event->command || $this->isCommandFiltered($event->command)) return;
+
+			$command = $this->app->make(ConsoleKernel::class)->all()[$event->command];
+
+			$argumentsDefaults = $command->getDefinition()->getArgumentDefaults();
+			$optionsDefaults = $command->getDefinition()->getOptionDefaults();
+
+			$this->app->make('clockwork')
+				->resolveAsCommand(
+					$event->command,
+					$event->exitCode,
+					array_diff($event->input->getArguments(), $argumentsDefaults),
+					array_diff($event->input->getOptions(), $optionsDefaults),
+					$argumentsDefaults,
+					$optionsDefaults,
+					$this->getConfig('artisan.collect_output') ? $event->output->getFormatter()->capturedOutput() : null
+				)
+				->storeRequest();
+		});
+	}
+
+	// Set up collecting of executed queue jobs
+	public function collectQueueJobs()
+	{
+		$this->app['events']->listen(\Illuminate\Queue\Events\JobProcessing::class, function ($event) {
+			// sync jobs are recorded as part of the parent request
+			if ($event->job instanceof \Illuminate\Queue\Jobs\SyncJob) return;
+
+			$payload = $event->job->payload();
+
+			if (! isset($payload['clockwork_id']) || $this->isQueueJobFiltered($payload['displayName'])) return;
+
+			$request = new Request([ 'id' => $payload['clockwork_id'] ]);
+			if (isset($payload['clockwork_parent_id'])) $request->setParent($payload['clockwork_parent_id']);
+
+			$this->app->make('clockwork')->reset()->setRequest($request)
+				->startEvent('total', 'Total execution time.', $request->time);
+		});
+
+		$this->app['events']->listen(\Illuminate\Queue\Events\JobProcessed::class, function ($event) {
+			$this->processQueueJob($event->job);
+		});
+
+		$this->app['events']->listen(\Illuminate\Queue\Events\JobFailed::class, function ($event) {
+			$this->processQueueJob($event->job, $event->exception);
+		});
+	}
+
+	protected function processQueueJob($job, $exception = null)
+	{
+		// sync jobs are recorded as part of the parent request
+		if ($job instanceof \Illuminate\Queue\Jobs\SyncJob) return;
+
+		$payload = $job->payload();
+
+		if (! isset($payload['clockwork_id'])) return;
+
+		$unserialized = isset($payload['data']['command']) ? unserialize($payload['data']['command']) : null;
+
+		if (! $unserialized || $this->isQueueJobFiltered(get_class($unserialized))) return;
+
+		if ($exception) {
+			$this->app->make('clockwork')->error($exception->getMessage(), [ 'exception' => $exception ]);
+		}
+
+		$this->app->make('clockwork')
+			->resolveAsQueueJob(
+				get_class($unserialized),
+				$payload['displayName'],
+				$job->hasFailed() ? 'failed' : ($job->isReleased() ? 'released' : 'done'),
+				$unserialized,
+				$job->getQueue(),
+				$job->getConnectionName(),
+				array_filter([
+					'maxTries'     => isset($payload['maxTries']) ? $payload['maxTries'] : null,
+					'delaySeconds' => isset($payload['delaySeconds']) ? $payload['delaySeconds'] : null,
+					'timeout'      => isset($payload['timeout']) ? $payload['timeout'] : null,
+					'timeoutAt'    => isset($payload['timeoutAt']) ? $payload['timeoutAt'] : null
+				])
+			)
+			->storeRequest();
+	}
+
 	public function process($request, $response)
 	{
 		if (! $this->isCollectingData()) {
@@ -186,9 +290,38 @@ class ClockworkSupport
 
 	public function isCollectingData()
 	{
+		return $this->isCollectingCommands()
+			|| $this->isCollectingQueueJobs()
+			|| $this->isCollectingRequests()
+			|| $this->isCollectingTests();
+	}
+
+	public function isCollectingCommands()
+	{
+		return ($this->isEnabled() || $this->getConfig('collect_data_always', false))
+			&& $this->app->runningInConsole()
+			&& $this->getConfig('artisan.collect', false);
+	}
+
+	public function isCollectingQueueJobs()
+	{
+		return ($this->isEnabled() || $this->getConfig('collect_data_always', false))
+			&& $this->app->runningInConsole()
+			&& $this->getConfig('queue.collect', false);
+	}
+
+	public function isCollectingRequests()
+	{
 		return ($this->isEnabled() || $this->getConfig('collect_data_always', false))
 			&& ! $this->app->runningInConsole()
 			&& ! $this->isUriFiltered($this->app['request']->getRequestUri());
+	}
+
+	public function isCollectingTests()
+	{
+		return ($this->isEnabled() || $this->getConfig('collect_data_always', false))
+			&& $this->app->runningInConsole()
+			&& $this->getConfig('tests.collect', false);
 	}
 
 	public function isFeatureEnabled($feature)
@@ -216,11 +349,6 @@ class ClockworkSupport
 		return $this->getConfig('web', true);
 	}
 
-	public function isWebUsingDarkTheme()
-	{
-		return $this->getConfig('web_dark_theme', false);
-	}
-
 	public function isUriFiltered($uri)
 	{
 		$filterUris = $this->getConfig('filter_uris', []);
@@ -235,10 +363,84 @@ class ClockworkSupport
 		return false;
 	}
 
+	protected function isCommandFiltered($command)
+	{
+		$whitelist = $this->getConfig('artisan.only', []);
+
+		if (count($whitelist)) return ! in_array($command, $whitelist);
+
+		$blacklist = $this->getConfig('artisan.except', []);
+
+		if ($this->getConfig('artisan.except_laravel_commands', true)) {
+			$blacklist = array_merge($blacklist, $this->builtinLaravelCommands());
+		}
+
+		$blacklist = array_merge($blacklist, $this->builtinClockworkCommands());
+
+		return in_array($command, $blacklist);
+	}
+
+	protected function isQueueJobFiltered($queueJob)
+	{
+		$whitelist = $this->getConfig('queue.only', []);
+
+		if (count($whitelist)) return ! in_array($queueJob, $whitelist);
+
+		$blacklist = $this->getConfig('queue.except', []);
+
+		return in_array($queueJob, $blacklist);
+	}
+
+	public function isTestFiltered($test)
+	{
+		$blacklist = $this->getConfig('tests.except', []);
+
+		return in_array($test, $blacklist);
+	}
+
 	protected function appendServerTimingHeader($response, $request)
 	{
 		if (($eventsCount = $this->getConfig('server_timing', 10)) !== false) {
 			$response->headers->set('Server-Timing', ServerTiming::fromRequest($request, $eventsCount)->value());
 		}
+	}
+
+	protected function builtinLaravelCommands()
+	{
+		return [
+			'clear-compiled', 'down', 'dump-server', 'env', 'help', 'list', 'migrate', 'optimize', 'preset', 'serve',
+			'tinker', 'up',
+			'app:name',
+			'auth:clear-resets',
+			'cache:clear', 'cache:forget', 'cache:table',
+			'config:cache', 'config:clear',
+			'db:seed',
+			'event:cache', 'event:clear', 'event:generate', 'event:list',
+			'key:generate',
+			'make:auth', 'make:channel', 'make:command', 'make:controller', 'make:event', 'make:exception',
+			'make:factory', 'make:job', 'make:listener', 'make:mail', 'make:middleware', 'make:migration', 'make:model',
+			'make:notification', 'make:observer', 'make:policy', 'make:provider', 'make:request', 'make:resource',
+			'make:rule', 'make:seeder', 'make:test',
+			'migrate:fresh', 'migrate:install', 'migrate:refresh', 'migrate:reset', 'migrate:rollback',
+			'migrate:status',
+			'notifications:table',
+			'optimize:clear',
+			'package:discover',
+			'queue:failed', 'queue:failed-table', 'queue:flush', 'queue:forget', 'queue:listen', 'queue:restart',
+			'queue:retry', 'queue:table', 'queue:work',
+			'route:cache', 'route:clear', 'route:list',
+			'schedule:run',
+			'session:table',
+			'storage:link',
+			'vendor:publish',
+			'view:cache', 'view:clear'
+		];
+	}
+
+	protected function builtinClockworkCommands()
+	{
+		return [
+			'clockwork:clean'
+		];
 	}
 }