+++ Clockwork/Support/Symfony/ClockworkConfiguration.php
@@ -2,6 +2,7 @@
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
+use Symfony\Component\HttpKernel\Kernel;
 
 class ClockworkConfiguration implements ConfigurationInterface
 {
@@ -14,14 +15,22 @@ class ClockworkConfiguration implements ConfigurationInterface
 
 	public function getConfigTreeBuilder()
 	{
-		return (new TreeBuilder)->root('clockwork')
+		return $this->getConfigRoot()
 			->children()
 				->booleanNode('enable')->defaultValue($this->debug)->end()
 				->booleanNode('web')->defaultValue(true)->end()
-				->booleanNode('web_dark_theme')->defaultValue(false)->end()
 				->booleanNode('authentication')->defaultValue(false)->end()
 				->scalarNode('authentication_password')->defaultValue('VerySecretPassword')->end()
 				->end()
 			->end();
+	}
+
+	protected function getConfigRoot()
+	{
+		if (Kernel::VERSION_ID < 40300) {
+			return (new TreeBuilder)->root('clockwork');
+		}
+
+		return (new TreeBuilder('clockwork'))->getRootNode();
 	}
 }