+++ Clockwork/Support/Symfony/ClockworkController.php
@@ -2,25 +2,25 @@
 
 use Clockwork\Clockwork;
 
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\Profiler\Profiler;
 
-class ClockworkController extends Controller
+class ClockworkController extends AbstractController
 {
 	protected $clockwork;
-	protected $profiler;
+	protected $support;
 
-	public function __construct(Clockwork $clockwork, ClockworkSupport $support, Profiler $profiler)
+	public function __construct(Clockwork $clockwork, ClockworkSupport $support)
 	{
 		$this->clockwork = $clockwork;
 		$this->support = $support;
-		$this->profiler = $profiler;
 	}
 
 	public function authenticate(Request $request)
 	{
+		$this->ensureClockworkIsEnabled();
+
 		$token = $this->clockwork->getAuthenticator()->attempt($request->request->all());
 
 		return new JsonResponse([ 'token' => $token ], $token ? 200 : 403);
@@ -28,33 +28,42 @@ class ClockworkController extends Controller
 
 	public function getData(Request $request, $id = null, $direction = null, $count = null)
 	{
-		$this->profiler->disable();
+		$this->ensureClockworkIsEnabled();
 
 		return $this->support->getData($request, $id, $direction, $count);
 	}
 
 	public function webIndex(Request $request)
 	{
-		$this->profiler->disable();
-
-		if ($this->support->isWebUsingDarkTheme() && ! $request->query->has('dark')) {
-			return $this->redirect('/__clockwork/app?dark');
-		}
+		$this->ensureClockworkIsEnabled();
+		$this->ensureClockworkWebIsEnabled();
 
 		return $this->support->getWebAsset('index.html');
 	}
 
 	public function webAsset($path)
 	{
-		$this->profiler->disable();
+		$this->ensureClockworkIsEnabled();
+		$this->ensureClockworkWebIsEnabled();
 
 		return $this->support->getWebAsset($path);
 	}
 
 	public function webRedirect()
 	{
-		$this->profiler->disable();
+		$this->ensureClockworkIsEnabled();
+		$this->ensureClockworkWebIsEnabled();
 
 		return $this->redirect('/__clockwork/app');
+	}
+
+	protected function ensureClockworkIsEnabled()
+	{
+		if (! $this->support->isEnabled()) throw $this->createNotFoundException();
+	}
+
+	protected function ensureClockworkWebIsEnabled()
+	{
+		if (! $this->support->isWebEnabled()) throw $this->createNotFoundException();
 	}
 }